+++ app/views/admin/tools/index.php
@@ -1,14 +1,5 @@
-<?php ob_start(); ?>
-<div class="row mb-2">
-    <div class="col-sm-6">
-        <h1 class="m-0">System Tools</h1>
-    </div>
-    <div class="col-sm-6">
-        <ol class="breadcrumb float-sm-right">
-            <li class="breadcrumb-item"><a href="/admin">Home</a></li>
-            <li class="breadcrumb-item active">System Tools</li>
-        </ol>
-    </div>
+</ol>
+</div>
 </div>
 
 <?php if (\App\Utils\RoleHelper::canImportExport()): ?>